fix: Renaming models
