+++ libs/api/domains/health-directorate/src/lib/models/appointmentDetail.model.ts
